--- NSGB/GridSearchBurst2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c9bda8aa195ea49b78df72b28e0eecb4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
